tests/wpt/web-platform-tests/workers/semantics/interface-objects/003.html

<!doctype html>
<title>available interface objects in shared worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
setup(function() {
  window.worker = new SharedWorker('003.js');
  worker.port.onmessage = function(e) {
    var result = e.data;
    for (var i = 0; i < result.length; ++i) {
      test(function() {
        assert_true(result[i][1]);
      }, "The " + result[i][0] + " interface object should be exposed");
    }
    done();
  }
}, {explicit_done: true});
</script>
